+++ nodes/Blinko/Blinko.node.ts
@@ -29,6 +29,10 @@ import { description } from './descriptions';
 // Import helpers
 import { getCredentialsAndBaseUrl } from './helpers';
 
+// Credentials
+import { BlinkoApiCredentials } from '../../credentials/BlinkoApi.credentials';
+
+
 export class Blinko implements INodeType {
 	description: INodeTypeDescription = {
 		displayName: 'Blinko',
@@ -45,7 +49,7 @@ export class Blinko implements INodeType {
 		outputs: [{ type: NodeConnectionType.Main }],
 		credentials: [
 			{
-				name: 'blinkoApi',
+				name: BlinkoApiCredentials.name,
 				required: true,
 			},
 		],